refactor(admin): better factorization of vivier components
