+++ app/organisms/CandidatePool/ApplicationSubtitle.tsx
@@ -1,0 +1,8 @@
+import { theme } from '@app/theme'
+import styled from 'styled-components'
+
+export const ApplicationSubtitle = styled.div`
+  font-size: 1.25rem;
+  line-height: 1.5rem;
+  color: ${theme.color.neutral.grey};
+`
